requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
Flask==3.0.3
Flask-SocketIO==5.3.6
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.15
langchain-community==0.2.14
langchain-core==0.2.36
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langsmith==0.1.107
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.43.0
orjson==3.10.7
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pystray==0.19.5
python-engineio==4.9.1
python-socketio==5.11.3
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.4
